Minor Bug Fixes
Fixed issue with extractor tool where trainer Pokemon's vanilla moves
were not extracted correctly.

Fixed issue where the late Rival fight could trigger repeatedly.

Fixed an issue with displaying the gym/badge count for NPCs who specify
those conditions.

+++ tools/extractor/extractor.h
@@ -93,7 +93,7 @@ class TrainerPokemonInfo {
     public:
         uint16_t species;
         uint8_t level;
-        uint8_t moves[4];
+        uint16_t moves[4];
 };
 
 class TrainerInfo {
